# Increase the #[inline] opportunities - 15-40% performance improvements