--- a/Cargo.toml
+++ b/Cargo.toml
@@ -24,7 +24,7 @@ no_std = [] # This is a no-op, preserved for backward compatibility only.
 
 [dev-dependencies]
 quickcheck = "0.7"
-bencher = "0.1"
+criterion = "0.3"
 
 [[bench]]
 name = "graphemes"