--- a/benches/graphemes.rs
+++ b/benches/graphemes.rs
@@ -1,55 +1,54 @@
-#[macro_use]
-extern crate bencher;
-extern crate unicode_segmentation;
+use criterion::{black_box, criterion_group, criterion_main, Criterion};
+use unicode_segmentation;
 
-use bencher::Bencher;
-use unicode_segmentation::UnicodeSegmentation;
 use std::fs;
+use unicode_segmentation::UnicodeSegmentation;
 
-fn graphemes(bench: &mut Bencher, path: &str) {
+fn graphemes(c: &mut Criterion, lang: &str, path: &str) {
     let text = fs::read_to_string(path).unwrap();
-    bench.iter(|| {
-        for g in UnicodeSegmentation::graphemes(&*text, true) {
-            bencher::black_box(g);
-        }
+
+    c.bench_function(&format!("grapheme {}",lang), |bench| {
+        bench.iter(|| {
+            for g in UnicodeSegmentation::graphemes(black_box(&*text), true) {
+                black_box(g);
+            }
+        })
     });
-
-    bench.bytes = text.len() as u64;
 }
 
-fn graphemes_arabic(bench: &mut Bencher) {
-    graphemes(bench, "benches/texts/arabic.txt");
+fn graphemes_arabic(c: &mut Criterion) {
+    graphemes(c, "arabic" ,"benches/texts/arabic.txt");
 }
 
-fn graphemes_english(bench: &mut Bencher) {
-    graphemes(bench, "benches/texts/english.txt");
+fn graphemes_english(c: &mut Criterion) {
+    graphemes(c, "english" ,"benches/texts/english.txt");
 }
 
-fn graphemes_hindi(bench: &mut Bencher) {
-    graphemes(bench, "benches/texts/hindi.txt");
+fn graphemes_hindi(c: &mut Criterion) {
+    graphemes(c, "hindi" ,"benches/texts/hindi.txt");
 }
 
-fn graphemes_japanese(bench: &mut Bencher) {
-    graphemes(bench, "benches/texts/japanese.txt");
+fn graphemes_japanese(c: &mut Criterion) {
+    graphemes(c, "japanese" ,"benches/texts/japanese.txt");
 }
 
-fn graphemes_korean(bench: &mut Bencher) {
-    graphemes(bench, "benches/texts/korean.txt");
+fn graphemes_korean(c: &mut Criterion) {
+    graphemes(c, "korean" ,"benches/texts/korean.txt");
 }
 
-fn graphemes_mandarin(bench: &mut Bencher) {
-    graphemes(bench, "benches/texts/mandarin.txt");
+fn graphemes_mandarin(c: &mut Criterion) {
+    graphemes(c, "mandarin" ,"benches/texts/mandarin.txt");
 }
 
-fn graphemes_russian(bench: &mut Bencher) {
-    graphemes(bench, "benches/texts/russian.txt");
+fn graphemes_russian(c: &mut Criterion) {
+    graphemes(c, "russian" ,"benches/texts/russian.txt");
 }
 
-fn graphemes_source_code(bench: &mut Bencher) {
-    graphemes(bench, "benches/texts/source_code.txt");
+fn graphemes_source_code(c: &mut Criterion) {
+    graphemes(c, "source_code","benches/texts/source_code.txt");
 }
 
-benchmark_group!(
+criterion_group!(
     benches,
     graphemes_arabic,
     graphemes_english,
@@ -61,4 +60,4 @@ benchmark_group!(
     graphemes_source_code,
 );
 
-benchmark_main!(benches);
+criterion_main!(benches);